Another Minor Update 4.0.1.1
yeah, masih soal long long int (karena ada overflow)

+++ Chalenge 1/header.h
@@ -27,9 +27,9 @@ void inputTanggal(int *hari, int *bulan, int *tahun, string username);
 int stdDate(int hari,int bulan, int tahun);
 
 //Input Data
-void dataInput(char *nama, long int *nomor, char *alamat, int *kodepos);
+void dataInput(char *nama, long long int *nomor, char *alamat, int *kodepos);
 
-void resiDisplay(string nama, long int nomor, string alamat, int kodepos , 
+void resiDisplay(string nama, long long int nomor, string alamat, int kodepos , 
                  string jenisIkan, string namaIkan, int kedalaman,
                  float harga, float diskon, float ongkir, float HargaTotal, float berat);
